chore: collapse helpers into main modules
